move searcher to search package
