+++ kifi-backend/modules/search/app/com/keepit/search/BookmarkSearcher.scala
@@ -5,7 +5,6 @@ import com.keepit.common.logging.Logging
 import com.keepit.common.time._
 import com.keepit.search.graph.URIGraphSearcherWithUser
 import com.keepit.search.index.DefaultAnalyzer
-import com.keepit.search.index.Searcher
 import com.keepit.search.query.SiteQuery
 import com.keepit.search.query.parser.DefaultSyntax
 import com.keepit.search.query.parser.QueryParser
