+++ kifi-backend/modules/search/app/com/keepit/search/query/ConditionalQuery.scala
@@ -1,7 +1,7 @@
 package com.keepit.search.query
 
 import com.keepit.common.logging.Logging
-import com.keepit.search.index.Searcher
+import com.keepit.search.Searcher
 import org.apache.lucene.index.AtomicReaderContext
 import org.apache.lucene.index.IndexReader
 import org.apache.lucene.index.Term
