+++ kifi-backend/modules/search/app/com/keepit/search/semantic/SemanticContextAnalyzer.scala
@@ -1,6 +1,6 @@
 package com.keepit.search.semantic
 
-import com.keepit.search.index.Searcher
+import com.keepit.search.Searcher
 import com.keepit.search.index.Analyzer
 import org.apache.lucene.analysis.tokenattributes.CharTermAttribute
 import java.io.StringReader
