+++ kifi-backend/modules/search/app/com/keepit/search/semantic/SemanticVectorSearcher.scala
@@ -3,7 +3,7 @@ package com.keepit.search.semantic
 import com.keepit.common.db.Id
 import com.keepit.model._
 import com.keepit.search.index.DefaultAnalyzer
-import com.keepit.search.index.Searcher
+import com.keepit.search.Searcher
 import com.keepit.search.graph.URIGraphSearcher
 import org.apache.lucene.analysis.tokenattributes.CharTermAttribute
 import org.apache.lucene.index.Term
